requirements.txt
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.4
kiwisolver==1.4.4
matplotlib==3.6.1
numpy==1.23.4
packaging==21.3
pacsltk==0.2.0
Pillow==9.2.0
plotly==5.11.0
PuLP==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
scipy==1.10.1
pandas
six==1.16.0
